:recycle: refactor: rename file selector to dom
