+++ src/js/view/IssuedTicketView.js
@@ -1,7 +1,7 @@
 import ticketTemplate from '../layouts/template.js';
 import { emit, on } from '../utils/event.js';
 import { $, $$ } from '../utils/selector.js';
-import ID from '../constants/selector.js';
+import ID from '../constants/dom.js';
 import EVENT from '../constants/event.js';
 
 /**
